+++ R/TheilSen.R
@@ -401,7 +401,7 @@ TheilSen <- function(mydata, pollutant = "nox", deseason = FALSE,
         ## s.window should not be "periodic"; set quite high to avoid
         ## overly fitted seasonal cycle
         ## robustness also makes sense for sometimes noisy data
-        ssd <- stl(myts, s.window = 35, robust = TRUE, s.degree = 0)
+        ssd <- stl(myts, s.window = 11, robust = TRUE, s.degree = 1)
         
         deseas <- ssd$time.series[, "trend"] + ssd$time.series[, "remainder"]